Improvments for tgrun cleanup
- reorginize cleaning functions
- cleanup secrets holding cloud-init data
