+++ testgrid/tgrun/pkg/runner/loop.go
@@ -17,7 +17,6 @@ import (
 	"k8s.io/client-go/kubernetes"
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
-	kubevirtv1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/client-go/kubecli"
 )
 
@@ -35,11 +34,11 @@ func MainRunLoop(runnerOptions types.RunnerOptions) error {
 	defer os.RemoveAll(tempDir)
 
 	for {
-		if err := CleanUpPVs(); err != nil {
-			fmt.Println("PV clean up ERROR: ", err)
-		}
 		if err := CleanUpVMIs(); err != nil {
 			fmt.Println("VMI clean up ERROR: ", err)
+		}
+		if err := CleanUpData(); err != nil {
+			fmt.Println("PV clean up ERROR: ", err)
 		}
 
 		canSchedule, err := canScheduleNewVM()
@@ -161,7 +160,7 @@ func GetRestConfig() (*restclient.Config, error) {
 	return config, nil
 }
 
-func GetClientset() (kubernetes.Interface, error) {
+func GetClientset() (*kubernetes.Clientset, error) {
 	config, err := GetRestConfig()
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to get restconfig")
@@ -194,92 +193,4 @@ func homeDir() string {
 		return h
 	}
 	return os.Getenv("USERPROFILE") // windows
-}
-
-// CleanUpPVs cleans stale PV/PVC and localpath to reclaim space
-func CleanUpPVs() error {
-	clientset, err := GetClientset()
-	if err != nil {
-		return errors.Wrap(err, "failed to get clientset")
-	}
-
-	pvcs, err := clientset.CoreV1().PersistentVolumeClaims(Namespace).List(metav1.ListOptions{})
-
-	// NOTE: OpenEbs webhook should be absent. For LocalPV volumes it has a bug preventing api calls, the webhook is only needed for cStor.
-	for _, pvc := range pvcs.Items {
-		// clean pvc older then 3 hours
-		if time.Since(pvc.CreationTimestamp.Time).Hours() > 3 {
-			pvc.ObjectMeta.SetFinalizers(nil)
-			p, _ := clientset.CoreV1().PersistentVolumeClaims(Namespace).Update(&pvc)
-			fmt.Printf("Removed finalizers on %s\n", p.Name)
-			clientset.CoreV1().PersistentVolumeClaims(Namespace).Delete(pvc.Name, &metav1.DeleteOptions{})
-			fmt.Printf("Deleted pvc %s\n", pvc.Name)
-		}
-	}
-
-	pvs, err := clientset.CoreV1().PersistentVolumes().List(metav1.ListOptions{})
-	if err != nil {
-		return errors.Wrap(err, "failed to get pv list")
-	}
-
-	// finalizers might get pv in a stack state
-	for _, pv := range pvs.Items {
-		localPath := pv.Spec.Local.Path
-		// deleting PVs older then 4 hours
-		if time.Since(pv.CreationTimestamp.Time).Hours() > 4 && pv.ObjectMeta.DeletionTimestamp == nil {
-			clientset.CoreV1().PersistentVolumes().Delete(pv.Name, &metav1.DeleteOptions{})
-			fmt.Printf("Deleted pv %s\n", pv.Name)
-			// Image file gets deleted and the space is reclamed on pv deletion
-			// However local directory is left with tmpimage file, removing
-			// 523M    /var/openebs/local/pvc-xxx
-			err := os.RemoveAll(localPath)
-			if err != nil {
-				fmt.Printf("Failed to delete %s; ERROR: %s", localPath, err)
-			}
-		} else if pv.ObjectMeta.DeletionTimestamp != nil {
-			// cleaning pv stack in Terminating state
-			pv.ObjectMeta.SetFinalizers(nil)
-			p, _ := clientset.CoreV1().PersistentVolumes().Update(&pv)
-			fmt.Printf("Removed finalizers on %s, localPath: %s\n", p.Name, localPath)
-		}
-	}
-	return nil
-}
-
-// CleanUpVMIs deletes "Succeeded" VMIs
-func CleanUpVMIs() error {
-	virtClient, err := GetKubevirtClientset()
-	if err != nil {
-		return errors.Wrap(err, "failed to get clientset")
-	}
-
-	vmiList, err := virtClient.VirtualMachineInstance(Namespace).List(&metav1.ListOptions{})
-	if err != nil {
-		return errors.Wrap(err, "failed to list vmis")
-	}
-
-	for _, vmi := range vmiList.Items {
-		// cleanup succeeded VMIs
-		// leaving them for a few hours for debug cases
-		if vmi.Status.Phase == kubevirtv1.Succeeded && time.Since(vmi.CreationTimestamp.Time).Hours() > 3 {
-			err := virtClient.VirtualMachineInstance(Namespace).Delete(vmi.Name, &metav1.DeleteOptions{})
-			if err != nil {
-				fmt.Printf("Failed to delete successful vmi %s: %v\n", vmi.Name, err)
-			} else {
-				fmt.Printf("Delete successful vmi %s\n", vmi.Name)
-			}
-		}
-
-		// cleanup VMIs that have been running for more than two hours
-		if vmi.Status.Phase == kubevirtv1.Running && time.Since(vmi.CreationTimestamp.Time).Minutes() > 120 {
-			err := virtClient.VirtualMachineInstance(Namespace).Delete(vmi.Name, &metav1.DeleteOptions{})
-			if err != nil {
-				fmt.Printf("Failed to delete long-running vmi %s: %v\n", vmi.Name, err)
-			} else {
-				fmt.Printf("Delete long-running vmi %s\n", vmi.Name)
-			}
-		}
-	}
-
-	return nil
 }
